bug fixes and improvement  - switched over to use chrono, instead of ctime for Timer class  - used WMV1 for video writer (smaller video size)  - used sleep_for, instead of Timer for FrameGrabber loop  - changed data type to match its usage

+++ src/include/EagleEye.hpp
@@ -39,7 +39,7 @@ private:
   Status status;
 
   int cam_id;
-  double record_sec;
+  int record_sec;
   myMat frame;
   FrameGrabber camFrameGrabber;
 };
